requirements.txt
frozenlist==1.4.1
cn2an==0.5.22
librosa==0.10.2.post1
opencc==1.1.7
pooch==1.8.1
torch==2.1.0+cu121
tbb==2021.11.0
torchaudio==2.1.0+cu121
reactivex==4.0.4
idna==3.4
packaging==24.0
ffmpeg==1.4
pypinyin==0.51.0
click==8.1.7
typing-extensions==4.11.0
pyaudio==0.2.14
annotated-types==0.6.0
unidecode==1.3.8
numpy==1.26.3
redis==5.0.4
torchvision==0.16.0+cu121
pyyaml==6.0.1
transformers==4.40.2
multidict==6.0.5
eng-to-ipa==0.0.2
ko-pron==1.3
websocket-client==1.8.0
typer==0.9.4
pyparsing==3.1.2
pillow==10.2.0
tokenizers==0.19.1
indic-transliteration==2.3.59
pydantic-core==2.18.2
charset-normalizer==2.1.1
antlr4-python3-runtime==4.9.3
pytz==2024.1
inflect==7.2.1
urllib3==1.26.13
pydantic==2.7.1

angchain
langchain-community
langchainhub
langchain-chroma
bs4

# 目前沒用到，但之後其他功能會用到先保留

# Cython==0.29.21
# matplotlib==3.3.1
# scikit-learn==1.0.2
# scipy==1.5.2
# tensorboard
# pyopenjtalk==0.1.3
# jamo==0.4.1
# jieba==0.42.1
# protobuf
# num_thai==0.0.5
# demucs
# openai-whisper
# gradio
# gy-redis
# whisper_live
